自己複製エージェント初期配置システムを実装
- AgentDefinition型と関連コードを削除
- エージェントプリセットシステムを新規実装
  - SingleHullSingleComputerAgentPreset型を定義
  - 自己複製エージェントプリセット（ダミープログラム付き）
  - AgentFactoryクラスでプリセットからエージェント生成
- World.addAgent()メソッドを追加
- WorldConfigにdefaultAgentPresetsプロパティを追加
- GameWorldで起動時に自己複製エージェントを配置

+++ src/engine/vm-executor.ts
@@ -803,7 +803,8 @@ export const InstructionExecutor = {
         const length = lengthRaw === 0 ? 256 : lengthRaw
         
         // メモリブロックをコピー
-        const memoryArray = vm.getMemoryArray() as Uint8Array
+        /* eslint-disable @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access */
+        const memoryArray = vm.getMemoryArray()
         const memorySize = memoryArray.length
         for (let i = 0; i < length; i++) {
           const srcIndex = (srcAddr + i) % memorySize
@@ -811,6 +812,7 @@ export const InstructionExecutor = {
           const value = memoryArray[srcIndex] ?? 0
           memoryArray[destIndex] = value
         }
+        /* eslint-enable @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access */
         
         vm.advancePC(decoded.length)
         return { success: true, cycles: 5 + length } // 基本5サイクル + バイト数
